# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Terramaps API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/maps": {
      "get": {
        "tags": [
          "Maps"
        ],
        "summary": "List Maps",
        "description": "List maps for the current user, each with its active job and import state.",
        "operationId": "list_maps_maps_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Map"
                  },
                  "type": "array",
                  "title": "Response List Maps Maps Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Maps"
        ],
        "summary": "Create Map",
        "description": "Create a map from a previously uploaded and parsed document.",
        "operationId": "create_map_maps_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMap"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/maps/{map_id}": {
      "patch": {
        "tags": [
          "Maps"
        ],
        "summary": "Patch Map",
        "description": "Rename a map. Any map member.",
        "operationId": "patch_map_maps__map_id__patch",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchMapDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Maps"
        ],
        "summary": "Get Map",
        "description": "Get a single map with its active job and import state.",
        "operationId": "get_map_maps__map_id__get",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/maps/{map_id}/exports/ppt": {
      "post": {
        "tags": [
          "PPT Exports"
        ],
        "summary": "Create Ppt Export",
        "description": "Create a new PPT export session.\n\nCancels any existing non-complete export for this map, then traverses the\nnode hierarchy to pre-compute all slide records. Bbox computation is deferred\nto GET /next so this returns immediately.",
        "operationId": "create_ppt_export_maps__map_id__exports_ppt_post",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/maps/{map_id}/exports/ppt/{export_id}/next": {
      "get": {
        "tags": [
          "PPT Exports"
        ],
        "summary": "Get Next Slide",
        "description": "Return the next slide capture instruction.\n\nFinds the lowest-order slide with no image_s3_key. Computes and persists its\nbbox on first access via PostGIS. Returns done=True once all slides have been\nuploaded.",
        "operationId": "get_next_slide_maps__map_id__exports_ppt__export_id__next_get",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextSlideResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/maps/{map_id}/exports/ppt/{export_id}/slides/{slide_id}": {
      "post": {
        "tags": [
          "PPT Exports"
        ],
        "summary": "Upload Slide",
        "description": "Upload the screenshot for a slide.\n\nStreams the image to S3 under a new UUID key and records it on the slide row.\nIdempotent: re-uploading replaces the previous S3 object.",
        "operationId": "upload_slide_maps__map_id__exports_ppt__export_id__slides__slide_id__post",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          },
          {
            "name": "slide_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Slide Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_slide_maps__map_id__exports_ppt__export_id__slides__slide_id__post"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/maps/{map_id}/exports/ppt/{export_id}/generate": {
      "post": {
        "tags": [
          "PPT Exports"
        ],
        "summary": "Generate Ppt",
        "description": "Enqueue the .pptx assembly task.\n\nVerifies all slides are uploaded, marks the export as 'generating',\ndispatches a Celery worker, and returns immediately. The client should\nthen poll GET /maps/{map_id}/exports/ppt/{export_id} for completion.\n\nIdempotent against re-entry while generating: returns the current state.",
        "operationId": "generate_ppt_maps__map_id__exports_ppt__export_id__generate_post",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/maps/{map_id}/exports/ppt/{export_id}": {
      "get": {
        "tags": [
          "PPT Exports"
        ],
        "summary": "Get Ppt Export Status",
        "description": "Return the current status of an export.\n\nWhen status='complete', includes a short-lived presigned URL for the .pptx.\nWhen status='failed', includes the error reason.",
        "operationId": "get_ppt_export_status_maps__map_id__exports_ppt__export_id__get",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PPT Exports"
        ],
        "summary": "Cancel Ppt Export",
        "description": "Cancel an export and delete all associated S3 files and DB rows.",
        "operationId": "cancel_ppt_export_maps__map_id__exports_ppt__export_id__delete",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_slide_maps__map_id__exports_ppt__export_id__slides__slide_id__post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_upload_slide_maps__map_id__exports_ppt__export_id__slides__slide_id__post"
      },
...
      "CreateExportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "total_slides": {
            "type": "integer",
            "title": "Total Slides"
          }
        },
        "type": "object",
        "required": [
          "id",
          "total_slides"
        ],
        "title": "CreateExportResponse",
        "description": "Returned immediately after a new export session is created."
      },
...
      "CreateMap": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/LayerSetup"
            },
            "type": "array",
            "title": "Layers"
          },
          "data_fields": {
            "items": {
              "$ref": "#/components/schemas/DataFieldSetup"
            },
            "type": "array",
            "title": "Data Fields"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "name",
          "layers",
          "data_fields"
        ],
        "title": "CreateMap",
        "description": "DTO for POST /maps.\n\nReferences a previously uploaded and parsed document. The raw spreadsheet\ndata stays in S3 and is read by the background import worker."
      },
      "DataFieldConfig": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": ""
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number"
            ],
            "title": "Type"
          },
          "aggregations": {
            "items": {
              "type": "string",
              "enum": [
                "sum",
                "avg",
                "min",
                "max"
              ]
            },
            "type": "array",
            "title": "Aggregations"
          },
          "precision": {
            "type": "integer",
            "title": "Precision",
            "default": 4
          }
        },
        "type": "object",
        "required": [
          "field",
          "type",
          "aggregations"
        ],
        "title": "DataFieldConfig",
        "description": "Data field config entry stored on a map."
      },
...
      "ExportStatusResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "generating",
              "complete",
              "failed"
            ],
            "title": "Status"
          },
          "total_slides": {
            "type": "integer",
            "title": "Total Slides"
          },
          "pptx_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pptx Url"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "total_slides"
        ],
        "title": "ExportStatusResponse",
        "description": "Returned by GET /maps/{map_id}/exports/ppt/{export_id}.\n\npptx_url and filename are populated only when status='complete';\nerror is populated only when status='failed'."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Map": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tile_version": {
            "type": "integer",
            "title": "Tile Version",
            "default": 0
          },
          "data_field_config": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DataFieldConfig"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Field Config"
          },
          "active_job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MapJob"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_export": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MapExport"
              },
              {
                "type": "null"
              }
            ]
          },
          "import_state": {
            "$ref": "#/components/schemas/MapImportState"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "import_state"
        ],
        "title": "Map",
        "description": "Map."
      },
...
      "MapJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "job_type": {
            "type": "string",
            "enum": [
              "recompute_geometry",
              "recompute_data"
            ],
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "complete",
              "failed"
            ],
            "title": "Status"
          },
          "step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_type",
          "status"
        ],
        "title": "MapJob",
        "description": "Background recompute job for a map (geometry or data). Import lifecycle is tracked separately via MapImportState."
      },
...
      "NextSlideResponse": {
        "properties": {
          "done": {
            "type": "boolean",
            "title": "Done"
          },
          "slide_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slide Id"
          },
          "order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "layer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layer Id"
          },
          "parent_node_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Node Id"
          },
          "bbox_min_lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox Min Lng"
          },
          "bbox_min_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox Min Lat"
          },
          "bbox_max_lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox Max Lng"
          },
          "bbox_max_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox Max Lat"
          },
          "total_slides": {
            "type": "integer",
            "title": "Total Slides"
          },
          "uploaded_slides": {
            "type": "integer",
            "title": "Uploaded Slides"
          }
        },
        "type": "object",
        "required": [
          "done",
          "total_slides",
          "uploaded_slides"
        ],
        "title": "NextSlideResponse",
        "description": "Returned by GET /next.\n\nWhen done=True all slide fields are None \u2014 the capture loop should stop.\nWhen done=False all slide fields are populated and the frontend should\ncapture + upload the described screenshot."
      },
...
      "PatchMapDTO": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PatchMapDTO"
      },
...
    }
  }
}